ui-test: isolate each tests
Isolate each test with a different `extest` process.
